+++ server/component/cms/tpl_new_ui/tpl_nav.php
@@ -22,7 +22,6 @@
                 <?php
                     if($_SESSION['logged_in'])
                     {
-                        $this->outputNavRight();
                         $this->output_profile();
                     }
                     else if($login)