+++ server/page/BasePage.php
@@ -365,8 +365,8 @@ abstract class BasePage
     {
         $csp_rules = "default-src 'self';  style-src 'self' 'unsafe-inline'; script-src 'self' 'unsafe-eval' 'sha256-"
             . base64_encode(hash('sha256', $this->get_js_constants(), true)) . "'; img-src 'self' blob: data: https://via.placeholder.com/;";
-        $csp_rules_plugins = $this->services->get_hooks()->getCspRules();
-        return $csp_rules . $csp_rules_plugins;
+        $csp_rules_hooks = $this->services->get_hooks()->getCspRules();
+        return $csp_rules . $csp_rules_hooks;
     }
 
     /**