+++ CHANGELOG.md
@@ -14,10 +14,9 @@
  - add option for page `is_headless` change
  - add markdown editor for fields from type `markdown`
  - add `hooks` which can be registered and executed in a plugin
-  - `outputNav` - Output an item in the navbar menu
-  - `outputNavRight` - Output an item in the right navbar menu
   - `outputStyleField` - Output a style filed based on its type
   - `getCspRules` - Output csp rules
+ - add `hooks` which can be added on function enter. They are stored in `hooks_onEnterFunction` - require module `php-uopz`
 
 ### Bugfix 
  - fix json condition builder for `in_one_of` type used with other condition